Refactored colour_detector to now take a Base64 encoded string rather than an image file
Moved this conversion into process_image.py to be used by command line scripts.
This enables the input to the function to be simpler, allowing for things such as a web UI that takes an image, base64 encodes the image and then submits that to a Lambda function rather than having to upload the image to something like S3 and then call the Lambda function with the location of the image to be processed.