Allow the use of Google Cloud Service Credentials
Checks for whether there is a non empty google-application-credentials.json file at the root of the repo and if so uses that as the credentials.
If the file is missing or empty then it defaults back to using user specified credentials (generated with the gcloud sdk as per the README).

To generate Service Account credentials follow the documentation here: https://cloud.google.com/vision/docs/common/auth
Alternatively you can still use the `gcloud beta auth application-default login` command to generate user credentials via a web flow when running locally.