Add API Gateway endpoint
Allows an HTTP post to an endpoint with a JSON object containing a base64 image and returns the closest product and the delta from the submitted image.
Decided to put a conditional block in so that it can be invoked raw for now as well as via the API gateway (that presents the base64_image inside a JSON document.
Might remove this at some point and just stick with the API gateway interface.